Fix issue with clicking open cards
Fix: Issue with clicking open cards

There was an issue with clicking cards more than once. When open cards were clicked the move counter would increment. Additionally, when the same card was clicked twice it would indicate it as a match.

Added, IDs to each card to be able to find it in the open cards array indicating it had already been clicked; in this situation we could ignore the click.